+++ src/plugins/clangcodemodel/clangfixitoperationsextractor.cpp
@@ -63,7 +63,7 @@ bool hasFixItAt(const QVector<ClangBackEnd::FixItContainer> &fixits,
 {
     const auto isFixitForLocation = [filePath, line] (const ClangBackEnd::FixItContainer &fixit) {
         const ClangBackEnd::SourceLocationContainer &location = fixit.range.start;
-        return location.filePath == filePath && location.line == uint(line);
+        return location.filePath == filePath && location.line == line;
     };
 
     return Utils::anyOf(fixits, isFixitForLocation);
